refactor: Rename files and delete unused files
