+++ src/vk/buffer.cpp
@@ -1,4 +1,4 @@
-#include "vk/vertex.h"
+#include "vk/buffer.h"
 
 #include <cstring>
